move non url related functions to helper file
