+++ exam_graph/views.py
@@ -34,76 +34,10 @@ def help(request):
     return HttpResponse('<h1>TODO: Add helpful tips for user!</h1>')
 
 def display_mock_csv(request):
-    df = build_test_master_json_df()
+    df = helper.build_test_master_json_df()
     graph = df.to_html()
     return render(request, 'test_template.html', {'graph': graph})
 
-def build_test_master_json_df() -> pd.DataFrame:
-
-    mock_json = pd.read_json('./mock_data.json')
-
-    return mock_json
-
-def parse_filter_request(request) -> dict: 
-
-    mock_json = build_test_master_json_df
-
-    if request.method == 'POST':
-        try:
-            # Decode and parse the JSON body
-            # "<class 'django.core.handlers.wsgi.WSGIRequest'>"
-            # the file uploaded by postman is a django.core.files.uploadedfile object
-            # we need to convert this object to bytes before we can json read it
-            if len(request.FILES) > 0:
-                # create InMemoryUploadedFile object from our mock_data.json stored in the 'test_file' key of our postman POST request 
-                in_memory_file:InMemoryUploadedFile = request.FILES['test_file']
-
-                # stackoverflow says to do this but idk what seek() does
-                in_memory_file.seek(0)
-
-                # read our object as bytes
-                file_bytes = in_memory_file.read()
-
-                # decode bytes to JSON string
-                file_string = file_bytes.decode('utf-8')
-
-                # convert from bytes to JSON file
-                file_json = json.loads(file_string)
-
-                # convert from JSON to pandas DataFrame
-                mock_json = pd.DataFrame.from_dict(file_json)
-
-
-            # parse form request
-
-            client_form = request.POST
-
-            metric = client_form['User_selected_metric']
-            # modality = [mod.strip() for mod in client_form['User_selected_modality'].split(',')] # this is for postman
-            modality = client_form.getlist('User_selected_modality')
-            period = client_form['period']
-            df = mock_json
-
-
-            post_req = {
-                'source dataframe': df,
-                'date range': '',
-                'xfilt': {
-                    'period': period,
-                    'modalities': modality
-                },
-                'User_selected_metric': metric,
-
-            }
-
-            return post_req
-        
-        except json.JSONDecodeError:
-            return JsonResponse({"error": "Invalid JSON body"}, status=400)
-
-    else:
-        return JsonResponse({"Your GET": request.method})
-    
 
 def upload_csv(request, modality):
     if request.method == 'POST' and request.FILES['csv_file']:
@@ -172,12 +106,12 @@ def upload_csv(request, modality):
 
 def filter_submission_handler(request):
 
-    parsed_mocked_data = build_test_master_json_df()
+    parsed_mocked_data = helper.build_test_master_json_df()
 
     try:
 
         # parse filter request
-        filter_params = parse_filter_request(request) # returns a dictionary containing the necessary arguments for master_filter()
+        filter_params = helper.parse_filter_request(request) # returns a dictionary containing the necessary arguments for master_filter()
 
         period = filter_params['xfilt']['period']
         modality_lst = filter_params['xfilt']['modalities']
